+++ src/main/java/dev/service/MissionService.java
@@ -10,6 +10,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import dev.Exception.MissionInvalidException;
 import dev.Exception.ModificationInvalideException;
 import dev.Utils.DtoUtils;
@@ -91,7 +92,7 @@ public class MissionService {
 		return missionList.stream().map(DtoUtils::toMissionDto).collect(Collectors.toList());
 	}
 
-	public List<MissionDto> recupererMissionParCollegue(Long id) {
+	public List<MissionDto> recupererMissionParCollegue(Integer id) {
 		Collegue collegue = collegueRepo.findById(id).orElseThrow(RuntimeException::new);
 		List<Mission> missionList = this.missionRepo.findByCollegue(collegue);
 		return missionList.stream().map(DtoUtils::toMissionDto).collect(Collectors.toList());
